requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# HTTP Client
httpx==0.27.0

# Data Validation
pydantic==2.9.0
pydantic-settings==2.5.0

# Environment Variables
python-dotenv==1.0.1

# Diff Parsing
unidiff==0.7.5

# LLM Integration - OpenAI + Google Generative AI
openai==1.54.0
langchain==0.3.7
langchain-openai==0.2.8
google-generativeai==0.8.3

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0

# Utilities
tenacity==9.0.0  # Retry logic